Improve the display of the list of winners.
Adding a table to display the regional ACervA, translating to Portuguese the BOS table and all the ordinals, and adding the total number of confirmed (judged) entries next to the BOS title. The paragraph just before the list of winners has more details about the number of participants, as well.

Another minor tweak was to delete the variable for the table "allowedcompetidores" from db_tables.inc.php since it is no longer used.